riscv: Rename riscv64 to riscv
`riscv64` is the arch name used for Rust, let change it to arch name
used in kernel source to align with other two architectures.
