+++ kvm-bindings/src/lib.rs
@@ -33,6 +33,6 @@ mod arm64;
 pub use self::arm64::*;
 
 #[cfg(target_arch = "riscv64")]
-mod riscv64;
+mod riscv;
 #[cfg(target_arch = "riscv64")]
-pub use self::riscv64::*;
+pub use self::riscv::*;
